Combined PANOC structures with impl
