+++ src/core/panoc/panoc_cache.rs
@@ -1,5 +1,34 @@
-use super::PANOCCache;
 use std::num::NonZeroUsize;
+
+/// Cache for PANOC
+///
+/// This struct carries all the information needed at every step of the algorithm.
+///
+/// An instance of `PANOCCache` needs to be allocated once and a (mutable) reference to it should
+/// be passed to instances of [PANOCEngine](struct.PANOCEngine.html)
+///
+/// Subsequently, a `PANOCEngine` is used to construct an instance of `PANOCAlgorithm`
+///
+#[derive(Debug)]
+pub struct PANOCCache {
+    pub(crate) lbfgs: lbfgs::Lbfgs,
+    pub(crate) gradient_u: Vec<f64>,
+    pub(crate) u_half_step: Vec<f64>,
+    pub(crate) gradient_step: Vec<f64>,
+    pub(crate) direction_lbfgs: Vec<f64>,
+    pub(crate) u_plus: Vec<f64>,
+    pub(crate) rhs_ls: f64,
+    pub(crate) lhs_ls: f64,
+    pub(crate) gamma_fpr: Vec<f64>,
+    pub(crate) gamma: f64,
+    pub(crate) tolerance: f64,
+    pub(crate) norm_gamma_fpr: f64,
+    pub(crate) tau: f64,
+    pub(crate) lipschitz_constant: f64,
+    pub(crate) sigma: f64,
+    pub(crate) cost_value: f64,
+    pub(crate) iteration: usize,
+}
 
 impl PANOCCache {
     /// Construct a new instance of `PANOCCache`
